Fix for "Unable to cast object of type 'System.String' to type 'Syste…

+++ src/API/LeadershipProfileAPI/Data/Models/StaffSearch.cs
@@ -8,7 +8,7 @@
         public string MiddleName { get; set; }
         public string LastSurname { get; set; }
         public string FullName { get; set; }
-        public int YearsOfService { get; set; }
+        public decimal YearsOfService { get; set; }
         public string Certification { get; set; }
         public string Assignment { get; set; }
         public string Degree { get; set; }
